+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 - Use MySQL script instead of Terraform provider to solve Terraform first run issue.
+- Refactor ECS task definition Environment variable names.
+- Migrate secrets from Hashicorp Vault to AWS SecretsManager.
 
 ### Added
 - Option to send Hive Metastore metrics to CloudWatch - see [#4](https://github.com/ExpediaInc/apiary-metastore-docker/issues/4).